Langchain/API/requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
langchain-openai
langserve
fastapi
uvicorn
langchain-community